requirements.txt
streamlit
speedtest-cli
pandas
plotly
requests
pyttsx3

